+++ src/buffer/scroll_view.rs
@@ -333,9 +333,15 @@ pub fn view<'a>(
                     + font::width_of_message_marker(&config.font);
         }
 
-        max_nick_width
+        let base = max_nick_width
             .max(range_end_timestamp_width)
-            .max(message_marker_width)
+            .max(message_marker_width);
+
+        if config.buffer.nickname.show_bot_icon {
+            base + theme::ICON_SIZE + 2.0
+        } else {
+            base
+        }
     });
 
     let max_prefix_width = max_prefix_chars