+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Added:
 - The default font is not bundled if default feature `iosevka-font` is disabled.
 - Configuration options for font stretch (`font.stretch`)
 - Notification for reacts
+- Bot mode support: a configurable icon is shown next to bots nicknames in buffer and nicklist
 
 Fixed: